server/src/interceptors/not-found.interceptor.ts
import {
  Injectable,
  NestInterceptor,
  ExecutionContext,
  NotFoundException,
  CallHandler,
} from "@nestjs/common";
import { Observable } from "rxjs";
import { tap } from "rxjs/operators";

// https://stackoverflow.com/a/60669415/4545255
@Injectable()
export class NotFoundInterceptor implements NestInterceptor {
  constructor(private errorMessage: string) {}

  intercept(context: ExecutionContext, stream$: CallHandler): Observable<any> {
    return stream$.handle().pipe(
      tap((data) => {
        if (data === null || data === undefined) {
          throw new NotFoundException(this.errorMessage);
        }
      }),
    );
  }
}
